backend_fastapi/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
numpy==1.24.4
pandas==2.1.4
joblib==1.3.2
python-dotenv==1.0.0
httpx==0.25.2
ta==0.10.2
scikit-learn==1.3.2
sqlmodel==0.0.14
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==7.4.3
pytest-asyncio==0.21.1
python-binance==1.0.19
cryptography==41.0.7